requirements.txt
customtkinter
python-docx
PyPDF2
pandas
openpyxl